Add explicit tests around (buggy?) behaviour introduced in Rails 3.0.2.
I'm yet to find the commit that introduced it but Rails 3.0.2 changes the behaviour of value_before_type_cast (see https://github.com/rails/rails/blob/v3.0.2/actionpack/lib/action_view/helpers/form_helper.rb#L1015) so that it tries to load the value from the database in favour of the value from the object instance.  This causes a problem with my methods (#date and #description) that are database columns that also have default values.

I plan to find the actual Rails commit and work out whether it was an intentional behaviour change.

For now, I've added the relevant duplication to ensure the form does the right thing and the tests pass.