A step toward more declarative features.
I've got some spiked code which changes the layout of the the transactions list.  The changes break quite a few of the features because they're specifically testing for a table.  I'm playing around with making them more declarative in the hope that it'll make them easier to fix when I introduce the layout changes.

*NOTE* I've overridden the Transaction.find_by_description dynamic finder so that I can find by the descriptions that are automatically generated.  There are two things I'm not entirely happy with here.  The first is that I have to check whether the description matches a name, memo and type or just a name and type.  The second is that I've made the search case-insensitive which the dynamic finders aren't.  Bah.