Visually indicate that transaction edits are being saved.
I wanted to have more confidence in the fact that the transaction details were being saved.  I think this is a good first step: highlight the cell on 'blur' and then gradually remove the highlight when we get a successful response from the server.

NOTE.  I have to remove the 'edited' class and 'backgroundColor' style when the fadeout completes successfully so that the element can be re-highlighted on a subsequent edit.

NOTE.  I'm using the jquery color plugin (http://plugins.jquery.com/project/color) but it looks like this might've been incorporated into the jquery UI plugin.