Users should easily be able to search for all transactions with a given description.
The most notable change is from the use of #within_transaction to using xpath in assert_debit_transaction.  This is because #within_transaction retrieved transactions based on their description yet, by definition, I needed to have consistent descriptions in the test I've added.  I think this approach is generally better because I'm checking what's on the page irrespective of implementation.  I will update assert_credit_transaction in a subsequent commit.