Bootstrapify the app
I've used the twitter-bootstrap-rails gem and ran `rails generate
bootstrap:install less` to generate the necessary scaffold.

I removed all my custom styles and then put a couple back in to make the table
look presentable.

I had to update the `xpath_category` helper method as it was checking that the
class attribute was a single value, and it's now 'category label label-info'.