Add autocomplete to category fields
This adds autocomplete to the category field in the bulk edit and single edit
forms.

I'm using jquery-ui-rails as suggested by the jquery-rails README.

I've added a class of 'category' to the input field in the bulk edit form and
single edit form, so that I can add the autocomplete functionality to only
those fields. I've set the minLength to 0 so that you can see all categories,
and the delay to 100 milliseconds so it feels responsive.

I've introduced a Category model, albeit possibly a little prematurely, as I
can imagine normalising the categories in the database.