Add Location to Transactions
This should allow me to split merchants from locations so that I can have, for
example, "Tesco" in both "City Road" and "Kingsland Road". I'm hoping this'll
allow me to get more accurate reporting of which merchants I spend most with,
and where those merchants are located.

I've switched to a fluid layout so that I have more room in the bulk edit form
- it now fits on my 13" laptop screen even with the new location field.

I've added a functional test around the updating of transactions, so that I
don't have to continue amending the cucumber scenarios. I'm imagining that
I'll start removing/simplifying the cucumber scenarios in the near future.

I've add two 'view tests' around the bulk edit and single edit forms. These
made it easier to add tests to ensure that I'd added the location field to the
correct forms.