Associate transactions with accounts
*NOTE* After running the migration in this commit, you'll want to update the
account_id of the account created, or manually merge the accounts (in the
Rails console) after a subsequent import.

* The migration creates a generic account and associates all existing
transactions with this account.

* The `StatementImporter` is responsible for finding or creating the account
referred to by the ACCTID field in the OFX file. It's possible I'll want to
restrict this further by BANKID in future but the ACCTID should be good enough
for now.

* Although accounts aren't referred to anywhere in the frontend, it *should*
(as of this commit) be possible to upload transactions from multiple banks and
not lose any data.