Fix a failing test
The addition of the 'should display original description in the edit form'
test in the previous commit caused this test to start failing. The test would
pass when run alone, or as part of all transaction controller tests, but it
would fail when run with rake. For whatever reason, the addition of the
aforementioned test resulted in `Transaction.search 'test description'`
returning the transactions in a different order.

Once I realised what was happening I focussed on my 'should add a class to the
first transaction on a given day' and noticed that it was asserting things
about the markup that it couldn't know for sure (i.e. the order in which the
transactions would be returned from `Transaction.search`). I've now updated it
so that it just checks for the presence of a single `firstOfDay` class.