Add autocomplete to the Location field
This is similar to the commits that added autocomplete to the
descriptions[1] and categories[2]. The difference here is that the set
of location options are dependent on the description field, in order to
limit the set of possibilities.

I wasn't able to use the simple format of `autocomplete` in
application.js, as I need to get hold of the description at the time I
make the request to get the set of autocomplete suggestions. Setting the
`source` option to a string doesn't allow me to do this.

I've added an additional 'transaction' class to the form element in the
edit template, so that I can target the transaction container given the
location field.

[1] d94e1bd7add127f773e6dd1c5c0972d34e6625f6
[2] c7a0ca0d14097eda9aad129a97e4e8182c32038f