Fix Rails/Output violations
There was a single violation in seeds.rb. I tried replacing the `puts` with
`warn` but that's not shown when you run the `db:seeds` rake task, which is the
purpose of this line. I'm not sure how else I could fix this so I'm just
excluding this file for this Cop.