Fix the missing icons
I needed to change back from Font Awesome to Glyphicons[1]. It wasn't quite as
simple as I'd hoped as the Glyphicons couldn't be found by the browser.

I initially set `@icon-font-path: '/assets/';` which meant the Glyphicons
displayed correctly in development but they didn't work in production.

I ended up having to redefine the `@font-face` so that I could use `font-url`
to have the icons display in both development and production.

[1]: https://github.com/seyhunak/twitter-bootstrap-rails#icons