Fix jquery-ui autocomplete styles
The autocomplete functionality was working but the list items were appearing on
a transparent background.

It looks like I introduced this problem in
419249e21239ffbccc0fcb91e576d0ead17988a7 when I changed from using `require` to
`@import` to load the autocomplete styles. I need to use `require` because
'jquery-ui/autocomplete' has additional `require` statements that all need to
be processed by Sprockets.

It looks like I may have been able to use the jquery-ui-sass-rails[1] Gem to
avoid having to revert to Sprockets, but I think my approach is a little
simpler.

[1]: https://github.com/jhilden/jquery-ui-sass-rails