Avoid ofx dependency git error on bundle install
The add-support-for-version-202-ofx-files branch on the chrisroos fork
has been deleted since this PR [1] was merged and `bundle install` was
failing with the following error:

    Fetching git://github.com/chrisroos/ofx.git
    fatal: Could not parse object '1b04543227d10da2d2100695071c142cc7f82184'.
    Git error: command `git reset --hard 1b04543227d10da2d2100695071c142cc7f82184` in directory /tmp/build_5c3996af94d35f3b545c5c1610e62590/vendor/bundle/ruby/2.2.0/bundler/gems/ofx-1b04543227d1 has failed.

I've chosen to manually set the SHA in the Gemfile.lock file to point at
this commit [2] which is the merge commit for the PR mentioned above.
This is only to reduce the scope of the change, not because I think
updating to the latest commit will necessarily be problematic.

[1]: https://github.com/annacruz/ofx/pull/14
[2]: https://github.com/annacruz/ofx/commit/70e99076440bd2915b10d59298c1d063b0ef6f57